refactor: rework service and user tokens components
