+++ frontend/components/dashboard/CliCommand.tsx
@@ -1,12 +1,14 @@
 import CopyButton from 'components/common/CopyButton'
 
-export const CliCommand = (props: { command: string; comment?: string }) => {
-  const prefixedCommand = `phase ${props.command}`
+export const CliCommand = (props: { command: string; comment?: string; prefix?: string }) => {
+  const prefix = props.prefix || 'phase'
+
+  const prefixedCommand = `${prefix} ${props.command}`
 
   return (
-    <div className="group relative overflow-x-auto rounded-lg border border-neutral-500/40 bg-zinc-300/50 dark:bg-zinc-800/50 p-4 text-left text-sm text-zinc-900 dark:text-zinc-100">
-      <pre>
-        <span className="text-emerald-800 dark:text-emerald-300">phase</span> {props.command}{' '}
+    <div className="group relative rounded-lg border border-neutral-500/40 bg-zinc-300/50 dark:bg-zinc-800/50 p-4 text-left text-sm text-zinc-900 dark:text-zinc-100">
+      <pre className="whitespace-pre-wrap">
+        <span className="text-emerald-800 dark:text-emerald-300">{prefix}</span> {props.command}{' '}
         {props.comment && (
           <span className="text-neutral-500">
             {'#'}
